pyproject.toml
[project]
name = "shiftlab"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["pyyaml"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
